Fixed #22 - Using sodium_memzero to remove temporary state variables created for GenericHash and PublicKeyAuth - Cleaning uup project and compiler warnings - Bumping vccorelib to vc141 - Sodium.winmd x86/Release is now provided in ref/uap10.0 instead of lib/uap10.0 - Added dependency on Microsoft.NETCore.UniversalWindowsPlatform >= 5.3.3 in Nuget to fix other packaging issues - Moving randombytes_implementation around, fixes ARM compilation issue

+++ libsodium-uwp/GenericHash.h
@@ -35,7 +35,7 @@ namespace Sodium
 			this->bytes = bytes;
 
 			crypto_generichash_state state;
-			size_t state_len = sizeof(state);
+			int state_len = sizeof(state);
 
 			int result = crypto_generichash_init(
 				&state,
